feat: set default wan interface as `auto` for auto detection
